+++ src/Components/Endpoints/src/Binding/Factories/Dictionary/TypedDictionaryConverterFactory.cs
@@ -151,118 +151,3 @@ internal class TypedDictionaryConverterFactory<TDictionaryType, TKey, TValue> : 
         throw new InvalidOperationException($"Unable to create converter for '{type.FullName}'.");
     }
 }
-
-internal class DictionaryBufferAdapter<TDictionaryType, TKey, TValue>
-    : IDictionaryBufferAdapter<TDictionaryType, TDictionaryType, TKey, TValue>
-    where TDictionaryType : IDictionary<TKey, TValue>, new()
-    where TKey : IParsable<TKey>
-{
-    public static TDictionaryType Add(ref TDictionaryType buffer, TKey key, TValue value)
-    {
-        buffer.Add(key, value);
-        return buffer;
-    }
-
-    public static TDictionaryType CreateBuffer() => new TDictionaryType();
-
-    public static TDictionaryType ToResult(TDictionaryType buffer) => buffer;
-}
-
-internal class ReadOnlyDictionaryBufferAdapter<TKey, TValue>
-    : IDictionaryBufferAdapter<ReadOnlyDictionary<TKey, TValue>, Dictionary<TKey, TValue>, TKey, TValue>
-    where TKey : IParsable<TKey>
-{
-    public static Dictionary<TKey, TValue> Add(ref Dictionary<TKey, TValue> buffer, TKey key, TValue value)
-    {
-        buffer.Add(key, value);
-        return buffer;
-    }
-
-    public static Dictionary<TKey, TValue> CreateBuffer() =>
-        new Dictionary<TKey, TValue>();
-
-    public static ReadOnlyDictionary<TKey, TValue> ToResult(Dictionary<TKey, TValue> buffer) =>
-        new ReadOnlyDictionary<TKey, TValue>(buffer);
-
-    internal static DictionaryConverter<IReadOnlyDictionary<TKey, TValue>> CreateInterfaceConverter(FormDataConverter<TValue> valueTypeConverter)
-    {
-        return new DictionaryConverter<IReadOnlyDictionary<TKey, TValue>,
-            DictionaryStaticCastAdapter<
-                IReadOnlyDictionary<TKey, TValue>,
-                ReadOnlyDictionary<TKey, TValue>,
-                ReadOnlyDictionaryBufferAdapter<TKey, TValue>,
-                Dictionary<TKey, TValue>,
-                TKey,
-                TValue>,
-            Dictionary<TKey, TValue>,
-            TKey,
-            TValue>(valueTypeConverter);
-    }
-
-    internal static DictionaryConverter<ReadOnlyDictionary<TKey, TValue>> CreateConverter(FormDataConverter<TValue> valueTypeConverter)
-    {
-        return new DictionaryConverter<ReadOnlyDictionary<TKey, TValue>,
-            ReadOnlyDictionaryBufferAdapter<TKey, TValue>,
-            Dictionary<TKey, TValue>,
-            TKey,
-            TValue>(valueTypeConverter);
-    }
-}
-
-internal class ImmutableDictionaryBufferAdapter<TKey, TValue>
-    : IDictionaryBufferAdapter<ImmutableDictionary<TKey, TValue>, ImmutableDictionary<TKey, TValue>.Builder, TKey, TValue>
-    where TKey : IParsable<TKey>
-{
-    public static ImmutableDictionary<TKey, TValue>.Builder Add(ref ImmutableDictionary<TKey, TValue>.Builder buffer, TKey key, TValue value)
-    {
-        buffer.Add(key, value);
-        return buffer;
-    }
-
-    public static ImmutableDictionary<TKey, TValue>.Builder CreateBuffer() => ImmutableDictionary.CreateBuilder<TKey, TValue>();
-
-    public static ImmutableDictionary<TKey, TValue> ToResult(ImmutableDictionary<TKey, TValue>.Builder buffer) => buffer.ToImmutable();
-
-    internal static DictionaryConverter<IImmutableDictionary<TKey, TValue>> CreateInterfaceConverter(FormDataConverter<TValue> valueTypeConverter)
-    {
-        return new DictionaryConverter<IImmutableDictionary<TKey, TValue>,
-            DictionaryStaticCastAdapter<
-                IImmutableDictionary<TKey, TValue>,
-                ImmutableDictionary<TKey, TValue>,
-                ImmutableDictionaryBufferAdapter<TKey, TValue>,
-                ImmutableDictionary<TKey, TValue>.Builder,
-                TKey,
-                TValue>,
-            ImmutableDictionary<TKey, TValue>.Builder,
-            TKey,
-            TValue>(valueTypeConverter);
-    }
-}
-
-internal class ImmutableSortedDictionaryBufferAdapter<TKey, TValue>
-    : IDictionaryBufferAdapter<ImmutableSortedDictionary<TKey, TValue>, ImmutableSortedDictionary<TKey, TValue>.Builder, TKey, TValue>
-    where TKey : IParsable<TKey>
-{
-    public static ImmutableSortedDictionary<TKey, TValue>.Builder Add(ref ImmutableSortedDictionary<TKey, TValue>.Builder buffer, TKey key, TValue value)
-    {
-        buffer.Add(key, value);
-        return buffer;
-    }
-
-    public static ImmutableSortedDictionary<TKey, TValue>.Builder CreateBuffer() => ImmutableSortedDictionary.CreateBuilder<TKey, TValue>();
-
-    public static ImmutableSortedDictionary<TKey, TValue> ToResult(ImmutableSortedDictionary<TKey, TValue>.Builder buffer) => buffer.ToImmutable();
-}
-
-internal class DictionaryStaticCastAdapter<TDictionaryInterface, TDictionaryImplementation, TDictionaryAdapter, TBuffer, TKey, TValue>
-    : IDictionaryBufferAdapter<TDictionaryInterface, TBuffer, TKey, TValue>
-    where TDictionaryAdapter : IDictionaryBufferAdapter<TDictionaryImplementation, TBuffer, TKey, TValue>
-    where TDictionaryImplementation : TDictionaryInterface
-    where TKey : IParsable<TKey>
-{
-    public static TBuffer CreateBuffer() => TDictionaryAdapter.CreateBuffer();
-
-    public static TBuffer Add(ref TBuffer buffer, TKey key, TValue element) => TDictionaryAdapter.Add(ref buffer, key, element);
-
-    public static TDictionaryInterface ToResult(TBuffer buffer) => TDictionaryAdapter.ToResult(buffer);
-}
